+++ core/src/main/java/org/eqasim/core/analysis/pt/PublicTransportLegWriter.java
@@ -6,15 +6,15 @@ import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.util.Collection;
 
-public class PublicTransportTripWriter {
-	final private Collection<PublicTransportTripItem> trips;
+public class PublicTransportLegWriter {
+	final private Collection<PublicTransportLegItem> trips;
 	final private String delimiter;
 
-	public PublicTransportTripWriter(Collection<PublicTransportTripItem> trips) {
+	public PublicTransportLegWriter(Collection<PublicTransportLegItem> trips) {
 		this(trips, ";");
 	}
 
-	public PublicTransportTripWriter(Collection<PublicTransportTripItem> trips, String delimiter) {
+	public PublicTransportLegWriter(Collection<PublicTransportLegItem> trips, String delimiter) {
 		this.trips = trips;
 		this.delimiter = delimiter;
 	}
@@ -25,7 +25,7 @@ public class PublicTransportTripWriter {
 		writer.write(formatHeader() + "\n");
 		writer.flush();
 
-		for (PublicTransportTripItem trip : trips) {
+		for (PublicTransportLegItem trip : trips) {
 			writer.write(formatTrip(trip) + "\n");
 			writer.flush();
 		}
@@ -38,21 +38,29 @@ public class PublicTransportTripWriter {
 		return String.join(delimiter, new String[] { //
 				"person_id", //
 				"person_trip_id", //
+				"leg_index", //
 				"access_stop_id", //
 				"egress_stop_id", //
 				"transit_line_id", //
-				"transit_route_id" //
+				"transit_route_id", //
+				"access_area_id", //
+				"egress_area_id", //
+				"transit_mode" //
 		});
 	}
 
-	private String formatTrip(PublicTransportTripItem trip) {
+	private String formatTrip(PublicTransportLegItem trip) {
 		return String.join(delimiter, new String[] { //
 				trip.personId.toString(), //
 				String.valueOf(trip.personTripId), //
+				String.valueOf(trip.legIndex), //
 				trip.accessStopId.toString(), //
 				trip.egressStopId.toString(), //
 				trip.transitLineId.toString(), //
-				trip.transitRouteId.toString() //
+				trip.transitRouteId.toString(), //
+				trip.accessAreaId == null ? "" : trip.accessAreaId.toString(), //
+				trip.egressAreaId == null ? "" : trip.egressAreaId.toString(), //
+				trip.transitMode //
 		});
 	}
 }