+++ core/src/main/java/org/eqasim/core/scenario/cutter/ScenarioWriter.java
@@ -42,6 +42,11 @@ public class ScenarioWriter {
 				.writeFile(new File(outputDirectory, prefix + "transit_schedule.xml.gz").toString());
 		new MatsimVehicleWriter(scenario.getTransitVehicles())
 				.writeFile(new File(outputDirectory, prefix + "transit_vehicles.xml.gz").toString());
+				
+		if (config.vehicles().getVehiclesFile() != null) {
+			new MatsimVehicleWriter(scenario.getVehicles())
+					.writeFile(new File(outputDirectory, prefix + "vehicles.xml.gz").toString());
+		}
 
 	}