+++ core/src/main/java/org/eqasim/core/simulation/analysis/EqasimAnalysisModule.java
@@ -2,10 +2,13 @@ package org.eqasim.core.simulation.analysis;
 
 import org.eqasim.core.analysis.DefaultPersonAnalysisFilter;
 import org.eqasim.core.analysis.PersonAnalysisFilter;
-import org.eqasim.core.analysis.TripListener;
+import org.eqasim.core.analysis.legs.LegListener;
+import org.eqasim.core.analysis.pt.PublicTransportLegListener;
+import org.eqasim.core.analysis.trips.TripListener;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.core.controler.AbstractModule;
 import org.matsim.core.router.MainModeIdentifier;
+import org.matsim.pt.transitSchedule.api.TransitSchedule;
 
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
@@ -25,4 +28,16 @@ public class EqasimAnalysisModule extends AbstractModule {
 		return new TripListener(network, mainModeIdentifier, personFilter);
 	}
 
+	@Provides
+	@Singleton
+	public LegListener provideLegListener(Network network, PersonAnalysisFilter personFilter) {
+		return new LegListener(network, personFilter);
+	}
+
+	@Provides
+	@Singleton
+	public PublicTransportLegListener providePublicTransportListener(Network network, TransitSchedule schedule,
+			PersonAnalysisFilter personFilter) {
+		return new PublicTransportLegListener(schedule);
+	}
 }