+++ core/src/main/java/org/eqasim/core/tools/routing/BatchPublicTransportRouter.java
@@ -4,11 +4,13 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.tuple.Pair;
 import org.eqasim.core.components.headway.HeadwayCalculator;
 import org.eqasim.core.misc.ParallelProgress;
 import org.matsim.api.core.v01.Coord;
-import org.matsim.api.core.v01.TransportMode;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.population.Leg;
 import org.matsim.core.network.NetworkUtils;
@@ -16,8 +18,12 @@ import org.matsim.core.router.LinkWrapperFacility;
 import org.matsim.facilities.Facility;
 import org.matsim.pt.router.TransitRouter;
 import org.matsim.pt.routes.TransitPassengerRoute;
+import org.matsim.pt.transitSchedule.api.Departure;
+import org.matsim.pt.transitSchedule.api.TransitLine;
+import org.matsim.pt.transitSchedule.api.TransitRoute;
 import org.matsim.pt.transitSchedule.api.TransitSchedule;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.inject.Provider;
 
 public class BatchPublicTransportRouter {
@@ -30,9 +36,11 @@ public class BatchPublicTransportRouter {
 	private final int numberOfThreads;
 	private final double interval;
 
+	private final Set<String> transitModes;
+
 	public BatchPublicTransportRouter(Provider<TransitRouter> routerProvider,
 			Provider<HeadwayCalculator> headwayCalculatorProvider, TransitSchedule schedule, Network network,
-			int batchSize, int numberOfThreads, double interval) {
+			int batchSize, int numberOfThreads, double interval, Set<String> transitModes) {
 		this.routerProvider = routerProvider;
 		this.headwayCalculatorProvider = headwayCalculatorProvider;
 		this.batchSize = batchSize;
@@ -40,11 +48,15 @@ public class BatchPublicTransportRouter {
 		this.schedule = schedule;
 		this.network = network;
 		this.interval = interval;
+		this.transitModes = transitModes;
 	}
 
-	public Collection<Result> run(Collection<Task> tasks) throws InterruptedException {
+	public Pair<Collection<TripInformation>, Collection<LegInformation>> run(Collection<Task> tasks)
+			throws InterruptedException {
 		Iterator<Task> taskIterator = tasks.iterator();
-		List<Result> results = new ArrayList<>(tasks.size());
+
+		List<TripInformation> tripResults = new ArrayList<>(tasks.size());
+		List<LegInformation> legResults = new ArrayList<>(tasks.size());
 
 		ParallelProgress progress = new ParallelProgress("Routing trips ...", tasks.size());
 		progress.start();
@@ -52,7 +64,7 @@ public class BatchPublicTransportRouter {
 		List<Thread> threads = new ArrayList<>(numberOfThreads);
 
 		for (int i = 0; i < numberOfThreads; i++) {
-			Thread thread = new Thread(new Worker(taskIterator, results, progress));
+			Thread thread = new Thread(new Worker(taskIterator, tripResults, legResults, progress));
 			threads.add(thread);
 			thread.start();
 		}
@@ -62,17 +74,20 @@ public class BatchPublicTransportRouter {
 		}
 
 		progress.close();
-		return results;
+		return Pair.of(tripResults, legResults);
 	}
 
 	private class Worker implements Runnable {
 		private final Iterator<Task> taskIterator;
-		private final Collection<Result> results;
+		private final Collection<TripInformation> tripResults;
+		private final Collection<LegInformation> legResults;
 		private final ParallelProgress progress;
 
-		private Worker(Iterator<Task> taskIterator, Collection<Result> results, ParallelProgress progress) {
+		private Worker(Iterator<Task> taskIterator, Collection<TripInformation> tripResults,
+				Collection<LegInformation> legResults, ParallelProgress progress) {
 			this.taskIterator = taskIterator;
-			this.results = results;
+			this.tripResults = tripResults;
+			this.legResults = legResults;
 			this.progress = progress;
 		}
 
@@ -94,10 +109,11 @@ public class BatchPublicTransportRouter {
 					}
 				}
 
-				List<Result> localResults = new ArrayList<>(localTasks.size());
+				List<TripInformation> localTripResults = new ArrayList<>(localTasks.size());
+				List<LegInformation> localLegResults = new ArrayList<>(localTasks.size() * 3);
 
 				for (Task task : localTasks) {
-					Result result = new Result(task);
+					TripInformation tripInformation = new TripInformation(task);
 
 					Coord fromCoord = new Coord(task.originX, task.originY);
 					Coord toCoord = new Coord(task.destinationX, task.destinationY);
@@ -109,13 +125,13 @@ public class BatchPublicTransportRouter {
 
 					if (legs != null) {
 						boolean isFirstVehicularLeg = true;
-						result.isOnlyWalk = 1;
+						tripInformation.isOnlyWalk = 1;
 
 						if (interval > 0.0) {
-							result.headway_min = headwayCalculator.calculateHeadway_min(fromFacility, toFacility,
-									task.departureTime);
+							tripInformation.headway_min = headwayCalculator.calculateHeadway_min(fromFacility,
+									toFacility, task.departureTime);
 						} else {
-							result.headway_min = Double.NaN;
+							tripInformation.headway_min = Double.NaN;
 						}
 
 						int currentIndex = 0;
@@ -123,143 +139,255 @@ public class BatchPublicTransportRouter {
 						for (Leg leg : legs) {
 							boolean isFirstLeg = currentIndex == 0;
 							boolean isLastLeg = currentIndex == legs.size() - 1;
-							currentIndex++;
 
-							if (leg.getMode().equals(TransportMode.access_walk)
-									|| (leg.getMode().equals(TransportMode.walk) && isFirstLeg)) {
-								result.accessTravelTime_min += leg.getTravelTime().seconds() / 60.0;
-								result.accessDistance_km += leg.getRoute().getDistance() * 1e-3;
-							} else if (leg.getMode().equals(TransportMode.egress_walk)
-									|| (leg.getMode().equals(TransportMode.walk) && isLastLeg)) {
-								result.egressTravelTime_min += leg.getTravelTime().seconds() / 60.0;
-								result.egressDistance_km += leg.getRoute().getDistance() * 1e-3;
-							} else if (leg.getMode().equals(TransportMode.transit_walk)
-									|| (leg.getMode().equals(TransportMode.walk) && !isFirstLeg && !isLastLeg)) {
-								result.transferTravelTime_min += leg.getTravelTime().seconds() / 60.0;
-								result.transferDistance_km += leg.getRoute().getDistance() * 1e-3;
-							} else if (leg.getMode().equals(TransportMode.pt)) {
+							if (leg.getMode().contains("walk") && isFirstLeg) {
+								tripInformation.accessTravelTime_min += leg.getTravelTime().seconds() / 60.0;
+								tripInformation.accessDistance_km += leg.getRoute().getDistance() * 1e-3;
+							} else if (leg.getMode().contains("walk") && isLastLeg) {
+								tripInformation.egressTravelTime_min += leg.getTravelTime().seconds() / 60.0;
+								tripInformation.egressDistance_km += leg.getRoute().getDistance() * 1e-3;
+							} else if (leg.getMode().contains("walk") && !isFirstLeg && !isLastLeg) {
+								tripInformation.transferTravelTime_min += leg.getTravelTime().seconds() / 60.0;
+								tripInformation.transferDistance_km += leg.getRoute().getDistance() * 1e-3;
+							} else if (transitModes.contains(leg.getMode())) {
 								TransitPassengerRoute route = (TransitPassengerRoute) leg.getRoute();
 
 								double waitingTime = route.getBoardingTime().seconds()
 										- leg.getDepartureTime().seconds();
 
 								if (isFirstVehicularLeg) {
-									result.initialWaitingTime_min += waitingTime / 60.0;
+									tripInformation.initialWaitingTime_min += waitingTime / 60.0;
 									isFirstVehicularLeg = false;
 								} else {
-									result.numberOfTransfers += 1;
-									result.transferWaitingTime_min += waitingTime / 60.0;
+									tripInformation.numberOfTransfers += 1;
+									tripInformation.transferWaitingTime_min += waitingTime / 60.0;
 								}
 
-								String mode = schedule.getTransitLines().get(route.getLineId()).getRoutes()
-										.get(route.getRouteId()).getTransportMode();
+								TransitLine transitLine = schedule.getTransitLines().get(route.getLineId());
+								TransitRoute transitRoute = transitLine.getRoutes().get(route.getRouteId());
+								String transitMode = transitRoute.getTransportMode();
 
 								double inVehicleTime = route.getTravelTime().seconds() - waitingTime;
 
-								switch (mode) {
+								switch (transitMode) {
 								case "rail":
-									result.inVehicleTimeRail_min += inVehicleTime / 60.0;
-									result.inVehicleDistanceRail_km += route.getDistance() * 1e-3;
+									tripInformation.inVehicleTimeRail_min += inVehicleTime / 60.0;
+									tripInformation.inVehicleDistanceRail_km += route.getDistance() * 1e-3;
 									break;
 								case "subway":
-									result.inVehicleTimeSubway_min += inVehicleTime / 60.0;
-									result.inVehicleDistanceSubway_km += route.getDistance() * 1e-3;
+									tripInformation.inVehicleTimeSubway_min += inVehicleTime / 60.0;
+									tripInformation.inVehicleDistanceSubway_km += route.getDistance() * 1e-3;
 									break;
 								case "bus":
-									result.inVehicleTimeBus_min += inVehicleTime / 60.0;
-									result.inVehicleDistanceBus_km += route.getDistance() * 1e-3;
+									tripInformation.inVehicleTimeBus_min += inVehicleTime / 60.0;
+									tripInformation.inVehicleDistanceBus_km += route.getDistance() * 1e-3;
 									break;
 								case "tram":
-									result.inVehicleTimeTram_min += inVehicleTime / 60.0;
-									result.inVehicleDistanceTram_km += route.getDistance() * 1e-3;
+									tripInformation.inVehicleTimeTram_min += inVehicleTime / 60.0;
+									tripInformation.inVehicleDistanceTram_km += route.getDistance() * 1e-3;
 									break;
 								default:
-									result.inVehicleTimeOther_min += inVehicleTime / 60.0;
-									result.inVehicleDistanceOther_km += route.getDistance() * 1e-3;
+									tripInformation.inVehicleTimeOther_min += inVehicleTime / 60.0;
+									tripInformation.inVehicleDistanceOther_km += route.getDistance() * 1e-3;
 								}
 
-								result.isOnlyWalk = 0;
+								tripInformation.isOnlyWalk = 0;
+
+								{ // Legs
+									Departure departure = findDeparture(route, transitRoute);
+
+									LegInformation legInformation = new LegInformation();
+									legInformation.identifier = task.identifier;
+									legInformation.legIndex = currentIndex;
+									legInformation.transitMode = transitMode;
+									legInformation.lineId = transitLine.getId().toString();
+									legInformation.routeId = transitRoute.getId().toString();
+									legInformation.vehicleId = departure.getVehicleId().toString();
+									legInformation.accessTime = route.getBoardingTime().seconds();
+									legInformation.egressTime = leg.getDepartureTime().seconds()
+											+ leg.getTravelTime().seconds();
+
+									localLegResults.add(legInformation);
+								}
 							} else {
 								throw new IllegalStateException();
 							}
+
+							currentIndex++;
 						}
 
-						result.inVehicleTimeTotal_min = result.inVehicleTimeRail_min + result.inVehicleTimeSubway_min
-								+ result.inVehicleTimeBus_min + result.inVehicleTimeTram_min
-								+ result.inVehicleTimeOther_min;
-						result.inVehicleDistanceTotal_km = result.inVehicleDistanceRail_km
-								+ result.inVehicleDistanceSubway_km + result.inVehicleDistanceBus_km
-								+ result.inVehicleDistanceTram_km + result.inVehicleDistanceOther_km;
-						result.totalWalkTravelTime_min = result.accessTravelTime_min + result.egressTravelTime_min
-								+ result.transferTravelTime_min;
-						result.totalWalkDistance_km = result.accessDistance_km + result.egressDistance_km
-								+ result.transferDistance_km;
+						tripInformation.inVehicleTimeTotal_min = tripInformation.inVehicleTimeRail_min
+								+ tripInformation.inVehicleTimeSubway_min + tripInformation.inVehicleTimeBus_min
+								+ tripInformation.inVehicleTimeTram_min + tripInformation.inVehicleTimeOther_min;
+						tripInformation.inVehicleDistanceTotal_km = tripInformation.inVehicleDistanceRail_km
+								+ tripInformation.inVehicleDistanceSubway_km + tripInformation.inVehicleDistanceBus_km
+								+ tripInformation.inVehicleDistanceTram_km + tripInformation.inVehicleDistanceOther_km;
+						tripInformation.totalWalkTravelTime_min = tripInformation.accessTravelTime_min
+								+ tripInformation.egressTravelTime_min + tripInformation.transferTravelTime_min;
+						tripInformation.totalWalkDistance_km = tripInformation.accessDistance_km
+								+ tripInformation.egressDistance_km + tripInformation.transferDistance_km;
 
-						localResults.add(result);
+						localTripResults.add(tripInformation);
 					}
 
 					progress.update();
 				}
 
-				synchronized (results) {
-					results.addAll(localResults);
+				synchronized (tripResults) {
+					tripResults.addAll(localTripResults);
+					legResults.addAll(localLegResults);
 				}
 			}
 		}
 	}
 
+	private static Departure findDeparture(TransitPassengerRoute passengerRoute, TransitRoute route) {
+		double boardingTime = passengerRoute.getBoardingTime().seconds();
+
+		List<Double> accessOffsets = route.getStops().stream() //
+				.filter(stop -> stop.getStopFacility().getId().equals(passengerRoute.getAccessStopId())) //
+				.map(stop -> stop.getArrivalOffset().seconds()).collect(Collectors.toList());
+
+		for (Departure departure : route.getDepartures().values()) {
+			if (departure.getDepartureTime() <= boardingTime) {
+				for (double offset : accessOffsets) {
+					if (departure.getDepartureTime() + offset == boardingTime) {
+						return departure;
+					}
+				}
+			}
+		}
+
+		throw new IllegalStateException("Departure not found");
+	}
+
 	static public class Task {
+		@JsonProperty("identifier")
 		public String identifier;
 
+		@JsonProperty("origin_x")
 		public double originX;
+
+		@JsonProperty("origin_y")
 		public double originY;
 
+		@JsonProperty("destination_x")
 		public double destinationX;
+
+		@JsonProperty("destination_y")
 		public double destinationY;
 
+		@JsonProperty("departure_time")
 		public double departureTime;
 	}
 
-	static public class Result {
+	static public class TripInformation {
+		@JsonProperty("identifier")
 		public String identifier;
 
+		@JsonProperty("access_travel_time_min")
 		public double accessTravelTime_min;
+
+		@JsonProperty("access_distance_km")
 		public double accessDistance_km;
 
+		@JsonProperty("egress_travel_time_min")
 		public double egressTravelTime_min;
+
+		@JsonProperty("egress_distance_km")
 		public double egressDistance_km;
 
+		@JsonProperty("transfer_travel_time_min")
 		public double transferTravelTime_min;
+
+		@JsonProperty("transfer_distance_km")
 		public double transferDistance_km;
 
+		@JsonProperty("total_walk_travel_time_min")
 		public double totalWalkTravelTime_min;
+
+		@JsonProperty("total_walk_distance_km")
 		public double totalWalkDistance_km;
 
+		@JsonProperty("in_vehicle_time_rail_min")
 		public double inVehicleTimeRail_min;
+
+		@JsonProperty("in_vehicle_time_subway_min")
 		public double inVehicleTimeSubway_min;
+
+		@JsonProperty("in_vehicle_time_bus_min")
 		public double inVehicleTimeBus_min;
+
+		@JsonProperty("in_vehicle_time_tram_min")
 		public double inVehicleTimeTram_min;
+
+		@JsonProperty("in_vehicle_time_other_min")
 		public double inVehicleTimeOther_min;
+
+		@JsonProperty("in_vehicle_time_total_min")
 		public double inVehicleTimeTotal_min;
 
+		@JsonProperty("in_vehicle_distance_rail_km")
 		public double inVehicleDistanceRail_km;
+
+		@JsonProperty("in_vehicle_distance_subway_km")
 		public double inVehicleDistanceSubway_km;
+
+		@JsonProperty("in_vehicle_distance_bus_km")
 		public double inVehicleDistanceBus_km;
+
+		@JsonProperty("in_vehicle_distance_tram_km")
 		public double inVehicleDistanceTram_km;
+
+		@JsonProperty("in_vehicle_distance_other_km")
 		public double inVehicleDistanceOther_km;
+
+		@JsonProperty("in_vehicle_distance_total_km")
 		public double inVehicleDistanceTotal_km;
 
+		@JsonProperty("transfers")
 		public int numberOfTransfers;
 
+		@JsonProperty("initial_waiting_time_min")
 		public double initialWaitingTime_min;
+
+		@JsonProperty("transfer_waiting_time_min")
 		public double transferWaitingTime_min;
 
+		@JsonProperty("headway_min")
 		public double headway_min;
 
+		@JsonProperty("is_only_walk")
 		public int isOnlyWalk;
 
-		Result(Task task) {
+		TripInformation(Task task) {
 			this.identifier = task.identifier;
 		}
+	}
+
+	static public class LegInformation {
+		@JsonProperty("identifier")
+		public String identifier;
+
+		@JsonProperty("leg_index")
+		public int legIndex;
+
+		@JsonProperty("line_id")
+		public String lineId;
+
+		@JsonProperty("route_id")
+		public String routeId;
+
+		@JsonProperty("vehicle_id")
+		public String vehicleId;
+
+		@JsonProperty("transit_mode")
+		public String transitMode;
+
+		@JsonProperty("access_time")
+		public double accessTime;
+
+		@JsonProperty("egress_time")
+		public double egressTime;
 	}
 }